[Fix] #3 - DetailTitleView 분리
